update slack notification
